Implement fix for buggy ASM1083 PCIe to PCI Bridge
On some machines, the root PCI bus is a PCI Express bus. PCI adaptors
can be plugged on the PCI Express bus thanks to a PCIe to PCI Bridge.

But the PCI and PCI Express worlds use a different method for interrupt
generation, so the PCIe/PCI Bridge needs to understand both methods.

Legacy PCI devices assert a dedicated pin to signal an interrupt,
whereas PCIe and modern PCI devices use MSI (Message Signaled
Interrupts) and don't have any interrupt pin.

When everything works fine, following happen:
1) The PCI device wishes to interrupt the host because an event occurred
and it asserts its interrupt pin.
2) When the PCI device's interrupt pin is asserted, the PCIe/PCI Bridge
understands this and generates a INTx_ASSERT message to the PCIe world.
3) Then, the message go to the I/O APIC (I/O Advanced Programmable
Interrupt Controller).
4) The I/O APIC decides to interrupt the CPU.
5) The CPU branches to the IRQ handler managed by the driver, the
handler deals with the interrupt and makes the PCI device stop to assert
its interrupt pin.
6) The PCIe/PCI Bridge sees that the PCI device's interrupt pin has been
de-asserted. It them generates a INTx_DEASSERT message to the PCIe
world.
7) The I/O APIC gets the message and stops interrupting the CPU.
8) The handler completes, the CPU returns from interrupt and resumes
normal operation.

The ASM1083 PCIe/PCI Bridge seems to have a bug where some INTx_DEASSERT
messages are lost and never reach the I/O APIC. This causes the I/O APIC
to keep interrupting the CPU even though the PCI device has nothing to
do and didn't assert its pin.
This causes our device's interrupt handler to be called a lot of times,
although the Interrupt Status Register shows the device didn't want to
interrupt us. In this case we return IRQ_NONE, since PCI IRQ lines can
be shared with other devices: it may just be that another device on the
same IRQ line is being interrupted.

In the Linux kernel, when within the last 100000 interrupts there have
been at least 99900 unhandled interrupts (all handlers returned
IRQ_NONE), it displays the message:
> irq 18: nobody cared (try booting with the "irqpoll" option)
Then the IRQ line is completely shut down. Depending on what other
device is sharing the interrupt line, this can be very harmful.
http://lxr.free-electrons.com/source/kernel/irq/spurious.c?v=3.16#L412

A preprocessor flag ASM1083_LOST_INTx_DEASSERT_FIX has been added. When
enabled, the driver will be compiled with a dirty fix which tries, when
IRQ_NONE occurs, to make the PCIe/PCI Bridge re-generate a new
INTx_DEASSERT message in the hope it will arrive to the I/O APIC which
will then finally stop the IRQ storm.  This is done simply by asking our
device to generate a useless interrupt, to be then able to clear it and
trigger another INTx_DEASSERT message generation. We do this using the
timer interruption mechanism of the RTL8139.

https://bugs.centos.org/view.php?id=6220
https://lkml.org/lkml/2012/1/30/216
https://bugzilla.kernel.org/show_bug.cgi?id=38632
https://lists.gt.net/linux/kernel/1461009